refactor: update `inquirer` to version `9.0.6`
Update the mentioned package.
