+++ package.json
@@ -99,7 +99,7 @@
     "@types/express": "^4.16.0",
     "@types/http-proxy": "^1.17.4",
     "@types/ini": "^1.3.31",
-    "@types/inquirer": "^8.0.0",
+    "@types/inquirer": "^9.0.6",
     "@types/jasmine": "~5.1.0",
     "@types/karma": "^6.3.0",
     "@types/less": "^3.0.3",
@@ -150,7 +150,7 @@
     "https-proxy-agent": "7.0.2",
     "husky": "8.0.3",
     "ini": "4.1.1",
-    "inquirer": "8.2.6",
+    "inquirer": "9.2.11",
     "jasmine": "^5.0.0",
     "jasmine-core": "~5.1.0",
     "jasmine-spec-reporter": "~7.0.0",
