+++ packages/angular/cli/package.json
@@ -29,7 +29,7 @@
     "@yarnpkg/lockfile": "1.1.0",
     "ansi-colors": "4.1.3",
     "ini": "4.1.1",
-    "inquirer": "8.2.6",
+    "inquirer": "9.2.11",
     "jsonc-parser": "3.2.0",
     "npm-package-arg": "11.0.1",
     "npm-pick-manifest": "9.0.0",
