+++ packages/angular_devkit/schematics_cli/package.json
@@ -19,7 +19,7 @@
     "@angular-devkit/core": "0.0.0-PLACEHOLDER",
     "@angular-devkit/schematics": "0.0.0-PLACEHOLDER",
     "ansi-colors": "4.1.3",
-    "inquirer": "8.2.6",
+    "inquirer": "9.2.11",
     "symbol-observable": "4.0.0",
     "yargs-parser": "21.1.1"
   }
